requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
langchain-community
langchain-openai
langchain-google-genai
PyPDF2
chromadb
faiss-cpu
pdf2image
db-sqlite3